breaking up RepositoryManagerDriverTest to make it parallelize better in the surefire plugin. Adding aprox promote client module, and wiring it in to promote build output (and also uncaptured dependencies) when persistArtifacts() is called from the maven repo driver. Adding in a collectionId field to the repository configuration, so it's easy to keep track the collection to which a build belongs.
