+++ maven-repository-manager/src/main/java/org/jboss/pnc/mavenrepositorymanager/RepositoryManagerDriver.java
@@ -2,7 +2,7 @@ package org.jboss.pnc.mavenrepositorymanager;
 
 import org.commonjava.aprox.client.core.Aprox;
 import org.commonjava.aprox.client.core.AproxClientException;
-import org.commonjava.aprox.client.core.util.UrlUtils;
+import org.commonjava.aprox.client.core.module.AproxContentClientModule;
 import org.commonjava.aprox.folo.client.AproxFoloAdminClientModule;
 import org.commonjava.aprox.folo.client.AproxFoloContentClientModule;
 import org.commonjava.aprox.folo.dto.TrackedContentDTO;
@@ -11,6 +11,9 @@ import org.commonjava.aprox.model.core.Group;
 import org.commonjava.aprox.model.core.HostedRepository;
 import org.commonjava.aprox.model.core.StoreKey;
 import org.commonjava.aprox.model.core.StoreType;
+import org.commonjava.aprox.promote.client.AproxPromoteClientModule;
+import org.commonjava.aprox.promote.model.PromoteRequest;
+import org.commonjava.aprox.promote.model.PromoteResult;
 import org.commonjava.maven.atlas.ident.ref.ArtifactRef;
 import org.commonjava.maven.atlas.ident.util.ArtifactPathInfo;
 import org.jboss.pnc.common.Configuration;
@@ -27,7 +30,10 @@ import org.jboss.pnc.spi.repositorymanager.model.RepositoryConfiguration;
 
 import java.io.File;
 import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
 import java.util.Properties;
 import java.util.Set;
 
@@ -53,16 +59,16 @@ public class RepositoryManagerDriver implements RepositoryManager {
 
     private static final String REPO_ID_FORMAT = "build+%s+%s";
 
-    private static final String PUBLIC_GROUP_ID = "public";
+    public static final String PUBLIC_GROUP_ID = "public";
 
-    private static final String SHARED_RELEASES_ID = "shared-releases";
+    public static final String SHARED_RELEASES_ID = "shared-releases";
 
-    private static final String SHARED_IMPORTS_ID = "shared-imports";
+    public static final String SHARED_IMPORTS_ID = "shared-imports";
 
     private Aprox aprox;
 
     @Deprecated
-    public RepositoryManagerDriver() { //workaround for CDI constructor parameter injection bug
+    public RepositoryManagerDriver() { // workaround for CDI constructor parameter injection bug
     }
 
     @Inject
@@ -70,7 +76,8 @@ public class RepositoryManagerDriver implements RepositoryManager {
         Properties properties = configuration.getModuleConfig(MAVEN_REPOSITORY_CONFIG_SECTION);
 
         String baseUrl = properties.getProperty(BASE_URL_PROPERTY);
-        aprox = new Aprox(baseUrl, new AproxFoloAdminClientModule(), new AproxFoloContentClientModule()).connect();
+        aprox = new Aprox(baseUrl, new AproxFoloAdminClientModule(), new AproxFoloContentClientModule(),
+                new AproxPromoteClientModule()).connect();
     }
 
     /**
@@ -124,23 +131,16 @@ public class RepositoryManagerDriver implements RepositoryManager {
         }
 
         // since we're setting up a group/hosted repo per build, we can pin the tracking ID to the build repo ID.
-        String url = trackingUrl(buildRepoId);
+        String url;
 
-        // TODO: Use this once we upgrade beyond AProx 0.17.0
-        // try {
-        // url = aprox.module(AproxFoloContentClientModule.class).trackingUrl(buildRepoId, StoreType.group, buildRepoId);
-        // } catch (AproxClientException e) {
-        // throw new RepositoryManagerException("Failed to retrieve AProx client module for the artifact tracker: %s", e,
-        // e.getMessage());
-        // }
+        try {
+            url = aprox.module(AproxFoloContentClientModule.class).trackingUrl(buildRepoId, StoreType.group, buildRepoId);
+        } catch (AproxClientException e) {
+            throw new RepositoryManagerException("Failed to retrieve AProx client module for the artifact tracker: %s", e,
+                    e.getMessage());
+        }
 
-        return new MavenRepositoryConfiguration(buildRepoId, new MavenRepositoryConnectionInfo(url));
-    }
-
-    // TODO: Remove once we upgrade beyond AProx 0.17.0
-    private String trackingUrl(String buildRepoId) {
-        return UrlUtils.buildUrl(aprox.getBaseUrl(), "/folo/track", buildRepoId, StoreType.group.singularEndpointName(),
-                buildRepoId);
+        return new MavenRepositoryConfiguration(buildRepoId, productRepoId, new MavenRepositoryConnectionInfo(url));
     }
 
     /**
@@ -150,25 +150,76 @@ public class RepositoryManagerDriver implements RepositoryManager {
      */
     @Override
     // TODO move under returned object (do not use the one from model) form createRepo
-    public void persistArtifacts(RepositoryConfiguration repository, BuildRecord buildResult)
-            throws RepositoryManagerException {
-        String trackingId = repository.getId();
+    public void persistArtifacts(RepositoryConfiguration repository, BuildRecord buildResult) throws RepositoryManagerException {
+        String buildId = repository.getId();
+
         TrackedContentDTO report;
         try {
-            report = aprox.module(AproxFoloAdminClientModule.class).getTrackingReport(trackingId, StoreType.group, trackingId);
+            report = aprox.module(AproxFoloAdminClientModule.class).getTrackingReport(buildId, StoreType.group, buildId);
         } catch (AproxClientException e) {
-            throw new RepositoryManagerException("Failed to retrieve tracking report for: %s. Reason: %s", e, trackingId,
+            throw new RepositoryManagerException("Failed to retrieve tracking report for: %s. Reason: %s", e, buildId,
                     e.getMessage());
+        }
+
+        processUploads(report, buildResult, repository);
+        processDownloads(report, buildResult);
+
+        // clean up.
+        try {
+            aprox.module(AproxFoloAdminClientModule.class).clearTrackingRecord(buildId, StoreType.group, buildId);
+            aprox.stores().delete(StoreType.group, buildId);
+            aprox.stores().delete(StoreType.remote, buildId);
+        } catch (AproxClientException e) {
+            throw new RepositoryManagerException(
+                    "Failed to clean up build repositories / tracking information for: %s. Reason: %s", e, buildId,
+                    e.getMessage());
+        }
+    }
+
+    /**
+     * Promote all build dependencies NOT ALREADY CAPTURED to the hosted repository holding store for the shared imports, then
+     * for each dependency artifact, add its metadata to the build result.
+     * 
+     * @param report The tracking report that contains info about artifacts downloaded by the build
+     * @param buildResult The build result where dependency artifact metadata should be appended
+     * @throws RepositoryManagerException In case of a client API transport error or an error during promotion of artifacts
+     */
+    private void processDownloads(TrackedContentDTO report, BuildRecord buildResult)
+            throws RepositoryManagerException {
+
+        AproxContentClientModule content;
+        try {
+            content = aprox.content();
+        } catch (AproxClientException e) {
+            throw new RepositoryManagerException("Failed to retrieve AProx client module. Reason: %s", e, e.getMessage());
         }
 
         Set<TrackedContentEntryDTO> downloads = report.getDownloads();
         if (downloads != null) {
             List<Artifact> deps = new ArrayList<>();
 
+            Map<StoreKey, Set<String>> toPromote = new HashMap<>();
+
+            StoreKey sharedImports = new StoreKey(StoreType.hosted, SHARED_IMPORTS_ID);
+            StoreKey sharedReleases = new StoreKey(StoreType.hosted, SHARED_RELEASES_ID);
+
             for (TrackedContentEntryDTO download : downloads) {
+                StoreKey sk = download.getStoreKey();
+                if (!sharedImports.equals(sk) && !sharedReleases.equals(sk)) {
+                    // this has not been captured, so promote it.
+                    Set<String> paths = toPromote.get(sk);
+                    if (paths == null) {
+                        paths = new HashSet<>();
+                        toPromote.put(sk, paths);
+                    }
+
+                    paths.add(download.getPath());
+                }
+
                 Artifact artifact = new Artifact();
                 artifact.setChecksum(download.getSha256());
-                artifact.setDeployUrl(download.getOriginUrl());
+
+                artifact.setDeployUrl(content.contentUrl(download.getStoreKey(), download.getPath()));
 
                 String path = download.getPath();
                 artifact.setFilename(new File(path).getName());
@@ -187,16 +238,51 @@ public class RepositoryManagerDriver implements RepositoryManager {
                 deps.add(artifact);
             }
 
+            for (Map.Entry<StoreKey, Set<String>> entry : toPromote.entrySet()) {
+                PromoteRequest req = new PromoteRequest(entry.getKey(), sharedImports, entry.getValue()).setPurgeSource(false);
+                doPromote(req);
+            }
+
             buildResult.setDependencies(deps);
         }
+    }
+
+    /**
+     * Promote all build output to the hosted repository holding store for the build collection to which this build belongs,
+     * then for each output artifact, add its metadata to the build result.
+     * 
+     * @param report The tracking report that contains info about artifacts uploaded (output) from the build
+     * @param buildResult The build result where output artifact metadata should be appended
+     * @param repository The AProx connection configuration containing the build- and collection-level repo id's
+     * @throws RepositoryManagerException In case of a client API transport error or an error during promotion of artifacts
+     */
+    private void processUploads(TrackedContentDTO report, BuildRecord buildResult, RepositoryConfiguration repository)
+            throws RepositoryManagerException {
+
+        AproxContentClientModule content;
+        try {
+            content = aprox.content();
+        } catch (AproxClientException e) {
+            throw new RepositoryManagerException("Failed to retrieve AProx client module. Reason: %s", e, e.getMessage());
+        }
+
+        String buildId = repository.getId();
+        String collectionId = repository.getCollectionId();
 
         Set<TrackedContentEntryDTO> uploads = report.getUploads();
         if (uploads != null) {
+            PromoteRequest promoteReq = new PromoteRequest(new StoreKey(StoreType.hosted, buildId), new StoreKey(
+                    StoreType.hosted, collectionId));
+
+            doPromote(promoteReq);
+
             List<Artifact> builds = new ArrayList<>();
 
             for (TrackedContentEntryDTO upload : uploads) {
                 Artifact artifact = new Artifact();
                 artifact.setChecksum(upload.getSha256());
+
+                content.contentUrl(StoreType.hosted, collectionId, upload.getPath());
                 artifact.setDeployUrl(upload.getLocalUrl());
 
                 String path = upload.getPath();
@@ -218,20 +304,46 @@ public class RepositoryManagerDriver implements RepositoryManager {
 
             buildResult.setBuiltArtifacts(builds);
         }
+    }
 
-        // NOTE: For the following, the artifacts need to be collated by GAV, since they are identified by GAVTC.
-        // TODO: Promote deps/downloads/imports (whatever we want to call them) into product-level (global?) imports hosted repo
-        // TODO: Promote uploads/build artifacts to product-level hosted repo.
-
-        // clean up.
+    /**
+     * Promotes a set of artifact paths (or everything, if the path-set is missing) from a particular AProx artifact store to
+     * another, and handle the various error conditions that may arise. If the promote call fails, attempt to rollback before
+     * throwing an exception.
+     * 
+     * @param req The promotion request to process, which contains source and target store keys, and (optionally) the set of
+     *        paths to promote
+     * @throws RepositoryManagerException When either the client API throws an exception due to something unexpected in
+     *         transport, or if the promotion process results in an error.
+     */
+    private void doPromote(PromoteRequest req) throws RepositoryManagerException {
+        AproxPromoteClientModule promoter;
         try {
-            aprox.module(AproxFoloAdminClientModule.class).clearTrackingRecord(trackingId, StoreType.group, trackingId);
-            aprox.stores().delete(StoreType.group, trackingId);
-            aprox.stores().delete(StoreType.remote, trackingId);
+            promoter = aprox.module(AproxPromoteClientModule.class);
         } catch (AproxClientException e) {
-            throw new RepositoryManagerException(
-                    "Failed to clean up build repositories / tracking information for: %s. Reason: %s", e, trackingId,
-                    e.getMessage());
+            throw new RepositoryManagerException("Failed to retrieve AProx client module. Reason: %s", e, e.getMessage());
+        }
+
+        try {
+            PromoteResult result = promoter.promote(req);
+            if (result.getError() != null) {
+                String addendum = "";
+                try {
+                    PromoteResult rollback = promoter.rollback(result);
+                    if (rollback.getError() != null) {
+                        addendum = "\nROLLBACK WARNING: Promotion rollback also failed! Reason given: " + result.getError();
+                    }
+
+                } catch (AproxClientException e) {
+                    throw new RepositoryManagerException("Rollback failed for promotion of: %s. Reason: %s", e, req,
+                            e.getMessage());
+                }
+
+                throw new RepositoryManagerException("Failed to promote: %s. Reason given was: %s%s", req, result.getError(),
+                        addendum);
+            }
+        } catch (AproxClientException e) {
+            throw new RepositoryManagerException("Failed to promote: %s. Reason: %s", e, req, e.getMessage());
         }
     }
 
@@ -348,6 +460,13 @@ public class RepositoryManagerDriver implements RepositoryManager {
      */
     private String safeUrlPart(String name) {
         return name.replaceAll("\\W+", "-").replaceAll("[|:]+", "-");
+    }
+
+    /**
+     * Convenience method for tests.
+     */
+    protected Aprox getAprox() {
+        return aprox;
     }
 
 }
