+++ maven-repository-manager/src/test/java/org/jboss/pnc/mavenrepositorymanager/RepositoryManagerDriver01Test.java
@@ -1,0 +1,41 @@
+package org.jboss.pnc.mavenrepositorymanager;
+
+import static org.hamcrest.CoreMatchers.equalTo;
+import static org.hamcrest.CoreMatchers.notNullValue;
+import static org.junit.Assert.assertThat;
+
+import org.jboss.pnc.model.BuildCollection;
+import org.jboss.pnc.model.BuildConfiguration;
+import org.jboss.pnc.spi.repositorymanager.model.RepositoryConfiguration;
+import org.jboss.pnc.spi.repositorymanager.model.RepositoryConnectionInfo;
+import org.junit.Test;
+
+public class RepositoryManagerDriver01Test 
+    extends AbstractRepositoryManagerDriverTest
+{
+
+    @Test
+    public void formatRepositoryURLForSimpleInfo_CheckDependencyURL() throws Exception {
+        BuildConfiguration pbc = simpleBuildConfiguration();
+
+        BuildCollection bc = new BuildCollection();
+        bc.setProductVersion(pbc.getProductVersion());
+
+        RepositoryConfiguration repositoryConfiguration = driver.createRepository(pbc, bc);
+
+        assertThat(repositoryConfiguration, notNullValue());
+
+        RepositoryConnectionInfo connectionInfo = repositoryConfiguration.getConnectionInfo();
+        assertThat(connectionInfo, notNullValue());
+
+        String expectedUrlPrefix = String.format("%sfolo/track/build+%s", url, pbc.getProject().getName());
+        String expectedGroupPathPrefix = String.format("/group/build+%s", pbc.getProject().getName());
+
+        assertThat("Expected URL prefix: " + expectedUrlPrefix + "\nActual URL was: " + connectionInfo.getDependencyUrl(),
+                connectionInfo.getDependencyUrl().startsWith(expectedUrlPrefix), equalTo(true));
+
+        assertThat("Expected URL to contain group path prefix: " + expectedGroupPathPrefix + "\nActual URL was: "
+                + connectionInfo.getDependencyUrl(), connectionInfo.getDependencyUrl().contains(expectedGroupPathPrefix),
+                equalTo(true));
+    }
+}
