+++ maven-repository-manager/src/test/java/org/jboss/pnc/mavenrepositorymanager/RepositoryManagerDriver02Test.java
@@ -1,0 +1,36 @@
+package org.jboss.pnc.mavenrepositorymanager;
+
+import static org.hamcrest.CoreMatchers.equalTo;
+import static org.hamcrest.CoreMatchers.notNullValue;
+import static org.junit.Assert.assertThat;
+
+import org.jboss.pnc.model.BuildCollection;
+import org.jboss.pnc.model.BuildConfiguration;
+import org.jboss.pnc.spi.repositorymanager.model.RepositoryConfiguration;
+import org.jboss.pnc.spi.repositorymanager.model.RepositoryConnectionInfo;
+import org.junit.Test;
+
+public class RepositoryManagerDriver02Test 
+    extends AbstractRepositoryManagerDriverTest
+{
+
+    @Test
+    public void formatRepositoryURLForSimpleInfo_AllURLsMatch() throws Exception {
+        BuildConfiguration pbc = simpleBuildConfiguration();
+
+        BuildCollection bc = new BuildCollection();
+        bc.setProductVersion(pbc.getProductVersion());
+
+        RepositoryConfiguration repositoryConfiguration = driver.createRepository(pbc, bc);
+        assertThat(repositoryConfiguration, notNullValue());
+
+        RepositoryConnectionInfo connectionInfo = repositoryConfiguration.getConnectionInfo();
+        assertThat(connectionInfo, notNullValue());
+
+        String expectedUrl = connectionInfo.getDependencyUrl();
+
+        assertThat(connectionInfo.getToolchainUrl(), equalTo(expectedUrl));
+        assertThat(connectionInfo.getDeployUrl(), equalTo(expectedUrl));
+    }
+
+}
