+++ maven-repository-manager/src/test/java/org/jboss/pnc/mavenrepositorymanager/fixture/Producer.java
@@ -1,4 +1,4 @@
-package org.jboss.pnc.mavenrepositorymanager;
+package org.jboss.pnc.mavenrepositorymanager.fixture;
 
 import org.commonjava.aprox.core.expire.ScheduleManager;
 import org.commonjava.aprox.inject.TestData;
