+++ pnc-spi/src/main/java/org/jboss/pnc/spi/repositorymanager/model/RepositoryConfiguration.java
@@ -11,5 +11,7 @@ public interface RepositoryConfiguration {
 
     String getId();
 
+    String getCollectionId();
+
     RepositoryConnectionInfo getConnectionInfo();
 }
