+++ pom.xml
@@ -45,7 +45,7 @@
 
   <properties>
     <atlasVersion>0.13.6</atlasVersion>
-    <aproxVersion>0.18.1</aproxVersion>
+    <aproxVersion>0.18.4-SNAPSHOT</aproxVersion>
     <version.jboss.maven.plugin>7.5.Final</version.jboss.maven.plugin>
     <version.jboss.bom>1.0.7.Final</version.jboss.bom>
     <version.jboss.as>7.2.0.Final</version.jboss.as>
@@ -361,6 +361,11 @@
         <scope>test</scope>
       </dependency>
 
+      <dependency>
+        <groupId>org.commonjava.aprox</groupId>
+        <artifactId>aprox-promote-client-java</artifactId>
+        <version>${aproxVersion}</version>
+      </dependency>
       <dependency>
         <groupId>org.commonjava.aprox</groupId>
         <artifactId>aprox-folo-client-java</artifactId>
